pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "calr-guard"
version = "0.3.1"
description = "Preflight validation and curation for calorimetry exports before CalR"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2",
  "openpyxl>=3.1",
  "gradio>=6.0",
  "pyyaml>=6.0",
  "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
calr-guard = "calr_guard.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["calr_guard*"]


[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
